+++ src/components/DreamTeam/Etiqueta.js
@@ -47,8 +47,9 @@ export default function Etiqueta(props) {
                 sx={{
                     pt: 0,
                     pb: 0,
+                    mr:0,
                     minWidth: "140px",
-                    maxWidth: "300px",
+                    maxWidth: "170px",
                     height: "40px",
                     borderRadius: "20px",
